+++ CHANGELOG.md
@@ -60,6 +60,8 @@ Modified clusters:
 - [matter.js]: Update to 0.13.0-alpha.0-20250405-7fc7db48.
 - [matter.js]: Update to 0.13.0-alpha.0-20250408-c916c7e8.
 - [matter.js]: Update to 0.13.0-alpha.0-20250412-5fad64e7b.
+- [matter.js]: Update to 0.13.0-alpha.0-20250413-d5a27700d.
+- [matter.js]: Update to 0.13.0-alpha.0-20250415-475996bb5.
 
 ### Fixed